Get -l flags from pkg-config (#31)
The pkg-config function already handles `-l` results from pkg-config. To
get them, we need to pass `--libs-only-l` to pkg-config as well. This
fixes a build error if you try to compile the library on environments
where tango libraries is not located in the same directory as the deps
of tango (zeromq and omniorb).