remove -h option to linker
This makes to sense. "-h FOO" is used to change the SONAME.
It comes right before "--strip-all". So the SONAME will be set
to "--strip-all".

This will again re-enable unconditional stripping of the _tango.so
file.

(This came up as an SONAME called "--strip-all" will cause problems
with tools that create rpm packages.)