Make __all__ definitions immutable
The flake8 linter complains (via pydocstyle) if `__all__` isn't
immutable, so changing to tuples.

This is the warning message:
```
WARNING: __all__ is defined as a list, this means pydocstyle cannot
reliably detect contents of the __all__ variable, because it can be
mutated. Change __all__ to be an (immutable) tuple, to remove this
warning. Note, pydocstyle uses __all__ to detect which definitions
are public, to warn if public definitions are missing docstrings.
If __all__ is a (mutable) list, pydocstyle cannot reliably assume
its contents. pydocstyle will proceed assuming __all__ is not mutated.
```