Add omni_thread::ensure_self to AutoPythonAllowThreads
This is to fix issues with multi-threaded applications, when using
"standard" threads, i.e. not omnithreads.  Typical failures occur
when using `DeviceProxy` objects created in different threads that
subscribe to events.

The cppTango layer assumes threads are omnithreads, and some of its
checks rely on them having an omnithread ID.  This `ensure_self`
object creates a dummy ID.

The `AutoPythonAllowThreads` class is a guard used to release the
GIL and protects all (most?) network I/O via the cppTango layer.
An `ensure_self` object will now exist for the lifetime of each
guard as well.