Add Docker development container
This adds a Dockerfile and some pinned environments that can be
used when developing PyTango.  This should simply the setup (especially
on non-Linux OSes), and allow for more consistent testing.  The most
useful part is being able to run and debug PyTango devices and tests
directly in an IDE.

The `.devcontainer` folder name is based on Visual Studio Code's
naming convention.