Add EnsureOmniThread and is_omni_thread
Subscription to events from non-omniORB threads can result in
application crashes.  A class has been added that exposes omniORB's
`omni_thread::ensure_self`.  This can be used as a context handler
in user threads to fix this problem.

Also added `is_omni_thread` which reports if a thread has an
omniORB thread ID or not.  This could be used in future to warn
or even raise an exception if methods that shouldn't be are called
from non-omniORB threads.  E.g., `DeviceProxy.subscribe_event`.