Generalise search for libboost_python on POSIX
The naming of the boost library has changed around version 1.67,
so the current approach in `setup.py` was failing on newer Linux
distributions.  Instead of hardcoding the suffix for different
distributions, use the `ctypes.utils.find_library` function and try
the various permutations.  The `find_library` function uses `ldconfig`
and a few other tools under the hood.

A new environment variable option was added for the case where
the search fails.  The library name can be specified directly by
setting `BOOST_LIB`.