Avoid Py3-only version of pyparsing on Py2
For some reason the CI tests for Python 2.7 are attempting to install
the pre-release version 3.0.0a1 of pyparsing.  This fails, as the
code is not compatible with Python 2.7.  Explicitly limit the version
to avoid this.  Note that pyparsing is not used directly by PyTango,
but rather a dependency of pytest.