Update AppVeyor CI with boost 1.73.0
For Python 3, use the newer 1.73.0 boost-ci release.  This will
also allow for Python 3.8 to be used in future.

Xamarin file deletion is not required - no warnings are showing up, so
remove this, as was done in the boost-ci repo.

Disabled the remote desktop (RDP) on AppVeyor failure as this just
wastes an hour per failed build.  Re-enable to debug a specific
problem, if necessary.