Fix DeviceProxy repr/str memory leak
The problem was caused by keeping the `info()` return
object in the `DeviceProxy` instance `__dict__`.

The boost layer is using the CallPolicy
`return_internal_reference<1>`
https://wiki.python.org/moin/boost.python/CallPolicy#return_internal_reference
This CallPolicy means that the `info()` return
object cannot be released until the `DeviceProxy`
instance is released.  However, the `DeviceProxy`
instance cannot be released until the `info()`
return object is released.  Python's garbage
collector cannot break this the cyclic reference
(probably because boost is involved).

The solution is to make a new object that copies
the fields from the original `info()` call, and
store that instead.  The existing `__TangoInfo`
was extended for this purpose.

It may be that changing the boost CallPolicy is
another way to fix it, but I don't know enough
about that to ensure no new problems are introduced.