Enable short name access to TestContext devices
When a device server is launched via the `(Multi)DeviceTestContext`,
it is run without a Tango database.  This means that access to any
of its devices must be done via a fully qualified domain name (FQDN).
I.e., we need something like `tango://172.17.0.3:48493/test/device/1#dbase=no`,
instead of just the short name: `test/device/1`.  Requiring the
FQDN is inconvenient.

This change stores the test device server's FQDN in an environment variable
while the TextContext is active.  Then, in order to make device
access with short names transparent, the `DeviceProxy`, `AttributeProxy`
and `Group` classes have been modified to take this environment variable
into account.  If it is set, and a short (i.e., unresolved) name is used
for the proxy constructor or group add method, then the name is rewritten
to a fully qualified (resolved) name.

Note that patterns (`*` wildcard) for groups are not supported.  Also, the
names of the items in the group will be FQDNs, rather than the short name
that was originally supplied.

There should not be a noticeable performance degradation for this outside
of test environments. It it just a single environment variable lookup
per proxy constructor or group add invocation.